Moved functions in test file
